Homestuck, more like eh-who-really-cares-stuck This commit may be reverted at a later date depending on what we decide our vision really is.
Pending a discussion between the mods, we've begun to reach the conclusion that Homestuck has shyed away from what we are seeking to create.
We are no longer actively seeking to directly emulate Homestuck. Rather, we want a modified survival server with some Homestuck-inspired mechanics and elements.

Machines and alchemy are great. Entry, the Medium, and Godtier as it is need work. We still want them as possiblities, but we won't be making obtaining access to them at all like the comic.