Woo, a language system. Eventually(TM) we'll move more stuff to it. Quite frankly, I'm just too tired of the refactor to bother right now.
Small-time stuff people won't be seeing much, admin-only commands, temporary stuff, etc. won't be migrated - the whole point of our command system is that it's fast, flexible, and simple. Maintaining two separate files for one bitty thing is not worth the bother, so we won't.

The real point of this language system is that I'm not going to maintain two separate forks for some tiny language adjustments for Ed, but it's silly to build a sweet little module like this and not use it for more things.