Initial 1.11 update
MAN, STUPID OLD API INCONSISTENCIES
MAN, STUPID NON-NULL ITEMSTACKS
MAN

This update was fiddly and annoying because it broke several bits of API-safe logic due to the fact that empty itemstacks are no longer null.

The old Bukkit API was horribly inconsistent; a lot of inventory events passed either null or an item with material air, so I suppose this change is good.